+++ CHANGELOG.md
@@ -5,7 +5,8 @@ next
 ----
 
 - Support of macro
-- Support of `{set}`, alias of `{assign}`
+- Support of `{set}`, an alias for `{assign}`
+- Support of `{verbatim}`, an alias for `{literal}`
 - Add `json_encode` modifier
 - Fix some exceptions
 - A little code cleanup